+++ gemrb/core/GUI/GameControl.h
@@ -236,7 +236,7 @@ public:
 	/** changes map to the current PC */
 	void ChangeMap(const Actor *pc, bool forced);
 	/** Sets up targeting with spells or items */
-	void SetupItemUse(int slot, int header, Actor *actor, int targettype, int cnt);
+	void SetupItemUse(int slot, size_t header, Actor *actor, int targettype, int cnt);
 	/** Page is the spell type + spell level info */
 	void SetupCasting(const ResRef& spellname, int type, int level, int slot, Actor *actor, int targettype, int cnt);
 	void ToggleAlwaysRun();
